Added CLI implementation, dry run and touched up on the result format

+++ main.py
@@ -1,8 +1,19 @@
-import os
+import os, sys
+import argparse
 import shutil
 import datetime
 
 folder = "/home/mahmoud/Github/Personal/automated-file-sorter/test_folder/"
+
+parser = argparse.ArgumentParser()
+
+parser.add_argument("--path", required = True)
+
+parser.add_argument("--mode", choices = ["ext", "date", "both"], required = True)
+
+parser.add_argument("--dry-run", action = "store_true")
+
+args = parser.parse_args()
 
 def list_files():
     for i, file in enumerate(files):
@@ -14,17 +25,16 @@ def file_definer(file):
     print(file_ext)
 
 def folder_generator(file):
-    if not os.path.isfile(os.path.join(folder, file)):w:
+    if not os.path.isfile(os.path.join(folder, file)):
         return
-    
+
     file_ext = os.path.splitext(file)[1]
     file_name = file_ext[1:] #Removes the dot in ".txt"
     full_path = os.path.join(folder, file_name)
     folder_created = os.makedirs(full_path, exist_ok = True)
     source = os.path.join(folder, file)
     destination = os.path.join(full_path, file)
-    shutil.move(source, destination)
-    print(f"Folder {file_name} made.")
+    move_file(source, destination, file, folder_name, file_name)
 
 def folder_generator_time(file):
     print("Checking:", file)
@@ -32,48 +42,48 @@ def folder_generator_time(file):
        return
     time_stamp = os.path.getmtime(os.path.join(folder, file))
     date = datetime.datetime.fromtimestamp(time_stamp)
-    folder_name = f"{date.year}-0{date.month}"
+    folder_name = f"{date.year}-0{date.month:02d}"
     full_path = os.path.join(folder, folder_name)
     folder_created = os.makedirs(full_path, exist_ok = True)
     source = os.path.join(folder, file)
     destination = os.path.join(folder, folder_name, file)
-    shutil.move(source, destination)
-    print(f"Moved {file} -> {folder_name}")
+    move_file(source, destination, file, folder_name, file_name)
 
 def folder_generator_all(file):
     if not os.path.isfile(os.path.join(folder, file)):
        return
     time_stamp = os.path.getmtime(os.path.join(folder, file))
     date = datetime.datetime.fromtimestamp(time_stamp)
-    folder_name = f"{date.year}-0{date.month}"
+    folder_name = f"{date.year}-0{date.month:02d}"
     file_ext = os.path.splitext(file)[1]
     file_name = file_ext[1:]
-    full_path = os.path.join(folder, file_name)
-    date_folder = os.makedirs(full_path, exist_ok = True)
-    ext_folder = os.makedirs(os.path.join(full_path, date_folder), exist_ok = True)
+    date_folder = folder_name
+    full_path = os.path.join(folder, file_name, date_folder)
+    ext_folder = os.makedirs(full_path, exist_ok = True)
     source = os.path.join(folder, file)
-    destination = os.path.join(folder, folder_name, file_name, file)
-    shutil.move(source, destination)
-    print(f"Moved {file} -> {folder_name}")
+    destination = os.path.join(full_path, file)
+    move_file(source, destination, file, folder_name, file_name)
     
+def move_file(source, destination, file, folder_name, file_name):
+    if args.dry_run:
+        print(f"[DRY RUN] {file} -> {file_name}/{folder_name}\n")
+    else:
+        shutil.move(source, destination)
+        print(f"Moved {file} -> {folder_name}/{file_name}")
 
 def main():
-    choice = input("File Sorting options (ext or date or dext) : ")
     
-    if choice == "ext":
-        for file in os.listdir(folder):
+    if args.mode == "ext":
+         for file in os.listdir(folder):
             folder_generator(file)
 
-    elif choice == "date":
+    elif args.mode == "date":
         for file in os.listdir(folder):
             folder_generator_time(file)
-    
-    elif choice == "dext":
-         for file in os.listdir(folder):
-             folder_generator_all(file)
-    else:
-        print("Invalid option")
 
+    elif args.mode == "both":
+        for file in os.listdir(folder):
+            folder_generator_all(file)
 
 if __name__ == "__main__":
     main()
